Attacking done
The environment now has a collidable feature that allows it to go through a generall assortment of collisions. Honestly this is an artifact but it might prove useful and I like the generalization.

The player class now has an Attacks list. When an attack is initiated, add it to the list. Attacks are their own classes to allow for weird things like freeze or smth, and deal with collisions on their own. Major thing here is that the attack class is now what does the sword animation not the player class. Attacks do need to be updated, but every attack in the list will be.